budget: colocate Budget and BudgetPeriod entities (#563)
Two budget-flavored entities consolidated to entities/budget.ts and
entities/budget-period.ts. Per-entity invariants (override sortedness,
periodStart < periodEnd) move next to the field declarations.
validateNoOverlappingPeriods stays at the collection level -- it
operates on BudgetPeriod[].

Refs #563.
