+++ budget/src/converters.ts
@@ -1,63 +1,31 @@
-import { Timestamp } from "firebase/firestore";
-import { DataIntegrityError } from "@commons-systems/firestoreutil/errors";
 import type {
-  Budget,
-  BudgetPeriod,
   Rule,
   NormalizationRule,
   WeeklyAggregate,
-  BudgetId,
-  BudgetPeriodId,
   RuleId,
   NormalizationRuleId,
-  GroupId,
-  AllowancePeriod,
 } from "./firestore.js";
-import type { IdbBudget, IdbBudgetPeriod, IdbRule, IdbNormalizationRule, IdbWeeklyAggregate } from "./idb.js";
+import { Timestamp } from "firebase/firestore";
+import type { IdbRule, IdbNormalizationRule, IdbWeeklyAggregate } from "./idb.js";
 import { idbToTransaction } from "./entities/transaction.js";
 import { idbToStatement } from "./entities/statement.js";
 import { idbToStatementItem } from "./entities/statement-item.js";
 import { idbToReconciliationNote } from "./entities/reconciliation-note.js";
+import { idbToBudget } from "./entities/budget.js";
+import type { Budget, IdbBudget } from "./entities/budget.js";
+import { idbToBudgetPeriod } from "./entities/budget-period.js";
+import type { BudgetPeriod, IdbBudgetPeriod } from "./entities/budget-period.js";
+import type { GroupId } from "@commons-systems/authutil/groups";
 
 export { idbToTransaction as toTransaction };
 export { idbToStatement as toStatement };
 export { idbToStatementItem as toStatementItem };
 export { idbToReconciliationNote as toReconciliationNote };
+export { idbToBudget as toBudget };
+export { idbToBudgetPeriod as toBudgetPeriod };
 
-function toAllowancePeriod(value: string | undefined): AllowancePeriod {
-  if (value === "monthly") return "monthly";
-  if (value === "quarterly") return "quarterly";
-  if (value == null || value === "weekly") return "weekly";
-  throw new DataIntegrityError(`Invalid allowancePeriod: ${value}`);
-}
-
-export function toBudget(row: IdbBudget): Budget {
-  return {
-    id: row.id as BudgetId,
-    name: row.name,
-    allowance: row.allowance,
-    allowancePeriod: toAllowancePeriod(row.allowancePeriod),
-    rollover: row.rollover,
-    overrides: (row.overrides ?? []).map(o => ({
-      date: Timestamp.fromMillis(o.dateMs),
-      balance: o.balance,
-    })),
-    groupId: null as GroupId | null,
-  };
-}
-
-export function toBudgetPeriod(row: IdbBudgetPeriod): BudgetPeriod {
-  return {
-    id: row.id as BudgetPeriodId,
-    budgetId: row.budgetId as BudgetId,
-    periodStart: Timestamp.fromMillis(row.periodStartMs),
-    periodEnd: Timestamp.fromMillis(row.periodEndMs),
-    total: row.total,
-    count: row.count,
-    categoryBreakdown: row.categoryBreakdown,
-    groupId: null as GroupId | null,
-  };
-}
+// Re-export types used by data-source.ts and other consumers
+export type { Budget, IdbBudget, BudgetPeriod, IdbBudgetPeriod };
 
 export function toRule(row: IdbRule): Rule {
   return {
